fix(bootstrap): increase depth for cloning submodules to reduce chance of error
With a depth of only `1`, the latest commit has to also be the one that
will be checked out, otherwise the checkout will fail. Increasing this
to `2` does allow some slight leeway in this regard, but `2` should
probably be the minimum.

So far I have found no better way than to arbitrarily select a value.
Therefore, I will start with `2` and increase as needed, if ever.