fix(chue): ensure symlink _directories_ are _not_ created
When changing colorschemes, if a directory that is unlinked does not
contain any other files, the next linkage (to the chosen colorscheme)
will mean that instead of symlinking only the _files_, stow will
actually symlink the entire directory instead.

This is a problem because with this method, the next linkage may not
work properly. Creating all directories (if needed) again just after the
unlinkage process fixes this.