fix(vim): not able to differentiate selected item from popup menu
This happened because the `pmenu` and `pmenusel` highlight groups were
the same or very similar colors.

This is easily fixed by referencing `buftablineactive` instead of
`pmenusel` from schemes as that highlight group only affects the
buftabline, and not the selected items of the popup menu, which is what
is desired.