feat(pentadactyl): remap back and forward to be more forceful
Normal back and forward actually functions like a vim jumplist, where
one back keybutton press might not actually go to the previous page, but
to the previous _position_ on the _same_ page.

This change overwrites such behaviour so that back and forward _always_
navigates the _history_ (instead of the jumplist).