fix(upd): check for updates in git repositories
Turns out checking for updatees is more complicated than simply checking
for the output of `git fetch --dry-run` because for whatever reason, a
shell script cannot interact with the output of that command.

However, we can compare the latest commit hash of the local repository
compared to the latest commit hash in the `master` branch of the remote
repository to do this. [1]

[1]: http://stackoverflow.com/a/15325376/4093264