feat(upd): switch to using output of `git status --porcelain` to check if working directory is clean or not
This allows to continue even if the working directory is unclean in
irrelevant areas:

If there is nothing staged for commit, and if there have been no
'meddling' with the files we are interested in (i.e. the
files/folders/paths that the script automatically adds and commits),
then the script continues.

This means untracked files, or unstaged changes to files the script does
not touch, are both not of concern to continue with the update process.
This is different to before, where even those circumstances would cause
the script to abort.