feat(vim): trim out extra not needed plugins
Not because they are not useful, they most definitely are. But it seems
as though Emacs will become the main editor, and so it makes sense to
trim the vim configuration down to the main essentials.

Vim shall still be installed alongside Emacs, just that it becomes
second choice as a 'basic' editor. *Oh how times have changed...*