feat(mudl): create new directory each time for download to happen in
This way each invocation of the script can run 'parallel' to all the
other invocations because everything is contained within separate
temporary directories.

Although this seems like a hacky way to do it, it seems to be a much
better way than the way the mdl script currently works, which is to
parse different lines of the same files as different downloads (which
obviously brings in a lot of edge cases and is therefore much harder to
manage).

Perhaps there is a better way to do this though...