feat(upd): parallelize the `git pull`s so that updating is *much* faster
Just like vim-plug does.

The one disadvantage though is that `git pull` does not output from
where it is pulling from (unlike for example with mercurial), so it is
not possible, or otherwise a bit difficult, to determine which
package/repository the corresponding output (status of the pull) belongs
to.

When the updates were synchronous, it was easy to determine this just by
running a simple `printf`/`echo` command before running `git pull`.
However, I think this should still be possible to solve by using a
post-merge hook (not sure), but to check definitively I have to wait for
an update to a package/repository...

On another note, if eventually I can make this update script much more
sophisticated in terms of detecting and showing the last changes, it may
be preferable to convert the vim plugins to git submodules as well, so
that there is no dichotomy between updating vim plugins versus all the
rest (except evil mode because that uses mercurial).