feat(qutebrowser): override some fonts to use default selections via css stylesheet
Of course, this means that attempting to use any of these fonts on *any*
site will result in it being replaced with the selections here.

Ideally this should only be done on particular sites as seen fit. This
can be fixed when qutebrowser gains the ability to have per-domain
settings [1].

[1]: https://github.com/The-Compiler/qutebrowser/issues/27