refactor(xmonad): remove default xmonad configuration from file listing
The reason this file was added in parent commit and now is subsequently
being removed with this change is because it feels more comfortable to
have the default configuration in the git history 'just in case'.

But of course, I don't need to keep the actual file in the directory!