fix(zsh): start mpd on startup
Cannot just put this in .zprofile as before because NixOS has to use a
graphical login manager right now (to use X), unfortunately. Meaning the
.zprofile does not get sourced. The most correct way to achieve this
probably to use systemd, however GuixSD does not use systemd (probably
also somewhat unfortunate), so in expectance to switch to GuixSD, that
will not be used.