feat(lemonbar): incorporate mpd information
There is a huge workaround/hack here where an actual file has to be used
that acts like a named pipe, instead of just using a pipe itself. -_-

For example, consider that `pipe` is a named pipe (i.e. made with
`mkfifo pipe`), running:

``` shell
while :; do date > pipe; sleep 1s; done
tail -f pipe
```

and then doing `mpc idleloop player > pipe` has no problems. Whenever,
an mpd player event happens, "player" is inputted into the pipe, and the
`tail -f pipe` shows that, *and* as soon as that is shown, you keep
seeing the time being displayed. i.e. the time input/output into the
pipe (by the `date` command above) is not blocked.

However, as you soon as you want to format the output of the pipe:

``` shell
while :; do date > pipe; sleep 1s; done
\# Consider that `formatter` is a script that formats the input to the
\# pipe (e.g. into something suitable for a panel program of some sort
\# to display).
tail -f pipe | formatter
```

and do `mpc idleloop player > pipe`, the whole output of the pipe is
blocked for whatever reason, until that mpc idleloop is ended.

Not sure if this a bug or intended behaviour, but the workaround
implemented here is to output the mpc idleloop output into a file, and
tail -f on that file, and then redirect that tail -f output into the
original pipe. This works and does not block the activity of the pipe,
even when piping the output of the pipe into a formatter.

Of course though, this a huge hack/limitation, not even sure why just
using a pipe doesn't work. Remember you also have to be careful that the
file does not grow, which is why the formatter script in this change
truncates the file on every player event input it gets (via `printf
''`). (Just using `>` instead of `>>` on the mpc idleloop line does not
seem to prevent the file from growing... what?)