fix(emacs/org): evil c* bindings do not work anymore
The previous "co" binding seemed to have overridden all the evil c*
bindings (which does not happen in vim, it works). It should still be
possible to have "co" bound to fi/org-collapse (because evil-surround
can bind to cs without messing up all the other c* bindings), but this
shall suffice for now at least. :/