feat(emacs/org): asynchronous export to pdf via (lua)latex
Asynchronous export actually spawns off another emacs intance to do the
export. Normally, `org-export-async-init-file` is set to the user's init
file, but my init file causes the export process to exit abnormally with
a strange error. Also, my init file does not actually load org (or ox),
thanks to use-package!

A 'cleaner' solution would be to parse the init file from (this) ox.el
and evaluate the :init form of the use-package form for org. That way,
all configuration still stays in the one file in a practical sense at
least.