refactor(mdl): do not mkdir-p the needed directories
mkdir in guile does not create non-existing parent directories. So
mkdir-p will have to do this, or the system "mkdir" will have to be
used, or we can simply assume that the user will have created all the
given directories beforehand.

The store status files are still created because those should solely be
used by mdl.