refactor(nixos): remove nixpkgs as submodule
Because it doesn't allow to do `git reset --hard
channels/nixos-unstable' from the submodule directory so it becomes
pretty useless. So have to use and manage a normal git repository
instead, which may even be better cause when a submodule is originally
cloned, it goes into a detached HEAD state, which iirc starts getting
weird when you're trying to manage the submodule from the parent repo.