feat(zsh): prompt
I quite like the arrow head, it makes it a lot easier to distinguish the
prompt line from command output. That being said, I don't know how well
it's going to look on other systems because that arrow head character
doesn't look good on my other systems (already tried it there). :|

So maybe I'll have to take it out anyway. Or maybe it has something to
do with different fonts on the other systems.