📝♻ readme: rewrite to make it simpler and focus on key elements only
I figured the most important things for a reader to immediately see on
the readme are screenshots, and an outline of what the repository aims
for.

It's hard to make an informed, correct decision, but because I predict
that the explanations and rationale for the aims will be quite long and
detailed, I've decided to only go with an outline, and delegate all that
extra detail into a separate documentation article. People who are
interested in it can just go there to reade more about it, whereas the
advantage is that the average visitor will not be greeted (and probably
scared off) by a huge wall of text.

This also gives more visibility to the Shoutouts section, which is
always nice.

The information removed here is going to be rewritten and re-added in
choices.org.