♻📝🎁 nixos: change installation process
Turns out trying to install the entire system from the live usb
environment has more complications than initially thought. Firstly, the
installer uses the /nix/store from the host system, so there probably
isn't enough space to download everything you would normally need for
your full installation.

That is a regression and even if that is fixed, the other limitation is
that this method (prior to this commit) of installation (i.e. via the -I
parameters to `nixos-install`), will not work if your version of nixpkgs
is older than the one that the live usb uses. And if you install with
the latest nixos-unstable image, then your nixpkgs has to be the latest
nixos-unstable version, which effectively eliminates the entire point of
pinning nixpkgs (if you always need to be on the latest version).

So instead, we install a minimal system from the new nixos/minimal.nix
file. And then finish the rest of the installation after rebooting into
that system.