⚙🎁 polybar: colors
The PBACKGROUND variable in x/resources needs to be defined like that
because simply using `#BACKGROUND` gets processed as "# fdfde7" instead
of the intended "#fdfde7". The space in between the the hex value and
the hash messes up the usage of the color value, so we have to end up
defining an entire new variable.

Not sure if that's a bug. 🤷‍♀️