🎁 polybar/xworkspaces
For showing x window workspaces. The label-occupied doesn't actually
seem to work, so this is a bit useless and our own script to output the
workspaces for herbstluftwm will have to be written. Also, if the
focused workspace gets a new urgent window, it will not show up as
urgent, so our own script has the potential to be slightly better.

Still, this is at least something for now.