⚙ emacs: do not use emojis in mode line
The emojis are great, but I hate how there's a bit of lag in emacs when
scrolling to the mode line configuration due the emojis there. Maybe
it's due to emacs handling of the font, or maybe it's due to bitmap
color emoji font format itself. I'll have to try with a different
SVGinOT color emoji font one day (i.e. not noto color emoji).

Also, textual (non-emoji) representation of evil state in mode line is
pretty useless imo when evil mode already displays '-- INSERT --' or '--
VISUAL --' in echo area (and don't forget different cursor shape), so
remove that part.